Move types into hook
