+++ context/app/static/js/components/workspaces/NewWorkspaceDialog/NewWorkspaceDialogFromSelections.tsx
@@ -11,8 +11,7 @@ import WorkspaceField from 'js/components/workspaces/WorkspaceField';
 import ErrorMessages from 'js/shared-styles/alerts/ErrorMessages';
 import useProtectedDatasetsForm from 'js/components/workspaces/NewWorkspaceDialog/useProtectedDatasetsForm';
 import { useHandleCopyClick } from 'js/hooks/useCopyText';
-import { useCreateWorkspaceForm } from './useCreateWorkspaceForm';
-import { CreateWorkspaceFormTypes } from './types';
+import { useCreateWorkspaceForm, CreateWorkspaceFormTypes } from './useCreateWorkspaceForm';
 
 function NewWorkspaceDialogFromSelections() {
   const { errorMessages, protectedHubmapIds, removeProtectedDatasets, protectedRows, selectedRows } =
