--- context/app/static/js/components/workspaces/NewWorkspaceDialog/types.ts
@@ -1,7 +1,0 @@
-interface CreateWorkspaceFormTypes {
-  'workspace-name': string;
-  'protected-datasets': undefined;
-  templates: string[];
-}
-
-export type { CreateWorkspaceFormTypes };
